Fix import for keys register_asset and move test to correct file
